#17 js파일을 assets로 추출후 파일을 불러오도록 변경
